Add distinction between modified and pristine cache
The schema for vmnetfs is changed to accomodate the modified cache path.
The modified cache is placed in the XDG_DATA_HOME directory, and the
cache info and path created in a matter similar to the pristine cache.
The existing cache variables are changed to distinguish between the
pristine and modified cache.